Fix test failures on conditional request (#299)
Instead of computing the value of the request header to use
(`If-Modified-Since` or `If-None-Match`), send a request with
no specific header, read the response header and reuse it in a
subsequent request to check that the server responds with a
`304 Not Modified` response. The test does not need to make the
header computation, just verify that the server can process it
as expected.

Fixes #299
