+++ controller/collaborators_blackbox_test.go
@@ -1,7 +1,6 @@
 package controller_test
 
 import (
-	"fmt"
 	"net/http"
 	"strings"
 	"testing"
@@ -18,7 +17,6 @@ import (
 	"github.com/fabric8-services/fabric8-auth/errors"
 	"github.com/fabric8-services/fabric8-auth/gormapplication"
 	"github.com/fabric8-services/fabric8-auth/gormtestsupport"
-	"github.com/fabric8-services/fabric8-auth/log"
 	"github.com/fabric8-services/fabric8-auth/resource"
 	"github.com/fabric8-services/fabric8-auth/space/authz"
 	testsupport "github.com/fabric8-services/fabric8-auth/test"
@@ -595,16 +593,6 @@ func assertResponseHeaders(t *testing.T, res http.ResponseWriter) (string, strin
 	cacheControl, err := getHeader(res, app.CacheControl)
 	require.NoError(t, err)
 	return *eTag, *lastModified, *cacheControl
-}
-
-func getHeader(res http.ResponseWriter, headerName string) (*string, error) {
-	values := res.Header()[headerName]
-	if len(values) == 0 {
-		return nil, fmt.Errorf("No '%s' header was found in the response", values)
-	}
-	value := values[0]
-	log.Debug(nil, map[string]interface{}{headerName: value}, "retrieved response header")
-	return &value, nil
 }
 
 type DummyResourceManager struct {
